issue: Client Side Column Sorting
This addresses an issue where client side column sorting does not work at
all. This is due to the if/else statement that checks for a REQUEST sort
order and if the REQUEST sort order matches an `$orderWays` array value. The
if statement returns TRUE for DESC and sets the sort order to DESC as it
equals '-' (a dash) but ASC equals '' (an empty string) so it returns FALSE
which fails-over to the else statement setting the sort order to DESC. In
addition, this adds sorting icons the the column headers to make it more
obvious they are sortable.