Referral Assignment Issue:
This commit fixes an issue where referrals were being considered successfully saved even if that referral already existed, resulting in multiple thread events for the same referral.

Instead, if we find that the thread is already referred to the referee, refer should return false so that an error message is displayed stating that we were unable to refer the ticket.