Task Assigned Team Issue:
This addresses an issue where it appeared as though tasks could not be assigned to teams if an agent in the team was not a part of the task's department. The assigment was actually taking place, however, it was not showing up in the Assignee column. Now the form will correctly display the column