Relative Time Fixes:
This commit removes the logic that was used for time within 2 days which would give a result of either today or tomorrow when something could actually be in 2 days instead.

It also rounds relative time that is within 29 days so that the result will be more accurate.
Ex: 1.99 will round to 2 instead of showing 1