queues: Counts revisited
Prefer agent's queue count instead of rough count when paginating the
tickets. This will make the initial queue load expensive but has an
added advantage of having queue counts available thereafter for drop downs.

This commits also adds entry to auto-cron, to keep queue counts more up to
date in the background.When APCu is not available SESSION is used to cache
the counts.