issue: ticket_link Fatal Error
This addresses an issue where replying to a ticket causes a Fatal Error for
the system. This is due to the `$ticket` object not being set before the
else statement therefore causing the `getId()` function to fail. This breaks
down the if/else statement to make sure the `$ticket` object is in-fact set
before continuing.