queue: Top-Level Ticket Counts
This adds a setting in the Ticket Settings to give the option of showing the
ticket counts on Top-Level queues. If enabled, this will populate the ticket
counts next to the Top-Level queue names like they used to be. The only
caveat is the counts are triggered by mousing-over any of the queue names as
we only load the counts when we need to display the sub-queues (for better
performance).

This also adds a new format function called `number` (Phase 1). As of now it
will format any integer to a comma separated number (eg. `1337 -> 1,337`).