issue: Email Default Dept
This addresses an issue where configuring a System Email to have a
Department and then editing/saving changes on the configured Department will
change the System Email's Department back to the the default Department.
This was due to a un-needed Department Status check introduced with commit
`1d361b5`. This also updates all instances of Dept Status comparison to be
case insensitive and non-translated.