issue: FAQ/Page Attachments
This addresses an issue where inserting images on a Page or FAQ and having
the setting enabled for "Login required to view attachments." does not
display the images at all when the Page/FAQ renders in view. This is due to
the Page/FAQ update functions that were not updated to fit the new
`id=>name` format for `keepOnlyFileIds()` introduced with `f179cf1`.