Issue: Mass Assign
This commit addresses an issue where a Department might have the Ticket Assignment option set to All Agents, but when trying to mass assign Tickets, only Department Agents were shown in the list.

Note: If mass assign is done on Tickets with different Departments that have different Ticket Assignment rules, Tickets will only be assigned to the Tickets that are allowed the assignment.