issue: New Ticket Field Permissions
This addresses an issue where creating a ticket as an agent and selecting a
Help Topic does not show the Issue Summary or Issue Details fields. This is
due to us passing 3 parameters instead of 1 as an array. This combines the
params in one array and only passes the array as a param.